+++ src/Telemetry/Microsoft.DotNet.AzureDevOpsTimelineTests/AzureDevOpsTimelineTests.cs
@@ -1,10 +1,6 @@
 using FluentAssertions;
 using Microsoft.DotNet.Internal.AzureDevOps;
 using Microsoft.DotNet.Internal.Testing.Utility;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Internal;
-using Microsoft.Extensions.Logging;
-using Moq;
 using NUnit.Framework;
 using System;
 using System.Collections.Generic;
@@ -16,205 +12,6 @@ namespace Microsoft.DotNet.AzureDevOpsTimeline.Tests
 {
     public class AzureDevOpsTimelineTests
     {
-        private sealed class TestData : IDisposable
-        {
-            public readonly AzureDevOpsTimeline Controller;
-            private readonly ServiceProvider _services;
-            public InMemoryTimelineTelemetryRepository Repository { get; }
-
-            public TestData(AzureDevOpsTimeline controller, InMemoryTimelineTelemetryRepository repository, ServiceProvider services)
-            {
-                Controller = controller;
-                Repository = repository;
-                _services = services;
-            }
-
-            public void Dispose()
-            {
-                _services?.Dispose();
-            }
-        }
-
-        private class TestDataBuilder
-        {
-            private readonly ServiceCollection collection = new ServiceCollection();
-            private InMemoryTimelineTelemetryRepository _repository = new InMemoryTimelineTelemetryRepository();
-
-            /// <summary>
-            /// Create a TestData builder with default configuration
-            /// </summary>
-            public TestDataBuilder()
-            {
-                collection.AddOptions();
-                collection.AddLogging(logging =>
-                {
-                    logging.AddProvider(new NUnitLogger());
-                });
-                collection.AddScoped<AzureDevOpsTimeline>();
-            }
-
-            public TestData Build()
-            {
-                collection.AddScoped<AzureDevOpsTimeline>();
-                collection.AddSingleton<ITimelineTelemetryRepository>(_repository);
-
-                ServiceProvider services = collection.BuildServiceProvider();
-                return new TestData(services.GetRequiredService<AzureDevOpsTimeline>(), _repository, services);
-            }
-
-            public TestDataBuilder AddBuildsWithTimelines(IEnumerable<BuildAndTimeline> builds)
-            {
-                var b = builds.ToDictionary(key => key.Build, key => key.Timelines.ToList());
-
-                collection.AddSingleton<IAzureDevOpsClient>(client => new MockAzureClient(b));
-
-                return this;
-            }
-
-            public TestDataBuilder AddTelemetryRepository()
-            {
-                return this;
-            }
-
-            public TestDataBuilder AddTelemetryRepository(string project, DateTimeOffset latestTime)
-            {
-                _repository = new InMemoryTimelineTelemetryRepository(
-                    new List<(string project, DateTimeOffset latestTime)> { (project, latestTime) });
-
-                return this;
-            }
-
-            public TestDataBuilder AddStaticClock(DateTimeOffset dateTime)
-            {
-                Mock<ISystemClock> mockSystemClock = new Mock<ISystemClock>();
-                mockSystemClock.Setup(x => x.UtcNow).Returns(dateTime);
-
-                collection.AddSingleton(mockSystemClock.Object);
-
-                return this;
-            }
-        }
-
-        public class BuildAndTimeline
-        {
-            public Build Build { get; }
-            public IList<Timeline> Timelines { get; } = new List<Timeline>();
-
-            public BuildAndTimeline(Build build)
-            {
-                Build = build;
-            }
-
-            public BuildAndTimeline(Build build, IList<Timeline> timelines)
-            {
-                Build = build;
-                Timelines = timelines;
-            }
-        }
-            
-        public class BuildAndTimelineBuilder
-        {
-            private BuildAndTimeline _build;            
-
-            public TimelineBuilder AddTimeline(string id, DateTimeOffset lastChangedOn)
-            {
-                return new TimelineBuilder(id, lastChangedOn);
-            }
-
-            public BuildAndTimelineBuilder AddTimeline(TimelineBuilder timelineBuilder)
-            {
-                _build.Timelines.Add(timelineBuilder.Build());
-
-                return this;
-            }
-
-            public BuildAndTimelineBuilder(Build build)
-            {
-                _build = new BuildAndTimeline(build);
-            }
-
-            public static BuildAndTimelineBuilder NewPullRequestBuild(int id, string projectName, string branchName)
-            {
-                Build build = new Build()
-                {
-                    Id = id,
-                    Project = new TeamProjectReference() { Name = projectName },
-                    ValidationResults = Array.Empty<BuildRequestValidationResult>(),
-                    Reason = "pullRequest",
-                    Parameters = $"{{\"system.pullRequest.targetBranch\": \"{branchName}\"}}"
-                };
-
-                return new BuildAndTimelineBuilder(build);
-            }
-
-            public BuildAndTimeline Build()
-            {
-                return _build;
-            }
-        }
-
-        public class TimelineBuilder
-        {
-            private Timeline timeline;
-            private List<TimelineRecord> records = new List<TimelineRecord>();
-
-            public Timeline Build()
-            {
-                timeline.Records = records.ToArray();
-                return timeline;
-            }
-
-            public TimelineBuilder(string id, DateTimeOffset lastChangedOn)
-            {
-                timeline = new Timeline()
-                {
-                    Id = id,
-                    LastChangedOn = lastChangedOn,
-                    Records = Array.Empty<TimelineRecord>()
-                };
-            }
-
-            public static TimelineBuilder EmptyTimeline(string id, DateTimeOffset lastChangedOn)
-            {
-                return new TimelineBuilder(id, lastChangedOn);
-            }
-
-            public TimelineBuilder AddRecord()
-            {
-                return AddRecord(null);
-            }
-
-            public TimelineBuilder AddRecord(string previousAttemptTimelineId)
-            {
-                int nextId = 1;
-                if (records.Any())
-                {
-                    nextId += records.Max(r => int.Parse(r.Id));
-                }
-
-                TimelineRecord record = new TimelineRecord()
-                {
-                    Id = nextId.ToString(),
-                    Issues = Array.Empty<TimelineIssue>()
-                };
-
-                if (string.IsNullOrEmpty(previousAttemptTimelineId))
-                {
-                    record.PreviousAttempts = Array.Empty<TimelineAttempt>();
-                }
-                else
-                {
-                    record.PreviousAttempts = new[] {
-                        new TimelineAttempt() { TimelineId = previousAttemptTimelineId }
-                    };
-                }
-
-                records.Add(record);
-
-                return this;
-            }
-        }
-
         /// <summary>
         /// Test a happy path. Database is empty, so all returned builds will be added.
         /// </summary>
